Modified README to help others run our code
Modified README to include link to the Object Detection code that we
used for Pepper. This repo contains a lot of pre-trained Object
Detection models. We couldn't successfully use the pre-trained models
for transfer learning due to lack of data.

This link is now given to help others to run our code smoothly, given
that we have passed out of NITT.

All the best for the upcoming years guys!
Signing Off.
Anirudh Swaminathan